requirements.txt
apache-airflow==2.7.1
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
h5py==3.9.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
requests==2.31.0

# FastAPI and API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6 